Last changes and fix some errors from product block

+++ app/Http/Controllers/Backend/ProductVariantController.php
@@ -94,10 +94,10 @@ class ProductVariantController extends Controller
     public function destroy(string $id)
     {
         $variant = ProductVariant::findOrFail($id);
-        //$variantItemCheck = ProductVariantItem::where('product_variant_id', $variant->id)->count();
-        //if($variantItemCheck > 0){
-         //   return response(['status' => 'error', 'message' => 'This variant contain variant items in it delete the variant items first for delete this variant!']);
-        //}
+        $variantItemCheck = ProductVariantItem::where('product_variant_id', $variant->id)->count();
+        if($variantItemCheck > 0){
+            return response(['status' => 'error', 'message' => 'This variant contain variant items in it delete the variant items first for delete this variant!']);
+        }
         $variant->delete();
 
         return response(['status' => 'success', 'message' => 'Deleted Successfully!']);
